Restructured app to use base classes
