+++ django_swagger_tester/__init__.py
@@ -1,6 +1,5 @@
-from django_swagger_tester.validate_response import validate_response  # noqa: F401
-from django_swagger_tester.validate_input import validate_input
+from django_swagger_tester import validate_response
 
 __version__ = '0.1.0'
 __author__ = 'Sondre Lillebø Gundersen'
-__all__ = ['validate_response', 'validate_input']
+__all__ = ['validate_response']
