+++ django_swagger_tester/validate_response/base/__init__.py
@@ -1,0 +1,1 @@
+from django_swagger_tester.validate_response.base.base import SwaggerTestBase
