+++ django_swagger_tester/validate_response/drf_yasg/__init__.py
@@ -1,0 +1,1 @@
+from django_swagger_tester.validate_response.drf_yasg.base import validate_response
