+++ django_swagger_tester/validate_response/static_schema/__init__.py
@@ -1,0 +1,1 @@
+from django_swagger_tester.validate_response.static_schema.base import validate_response
